Fix DeltePoseWithDeltaTime.msg
duration has been replaced by builtin_interfaces/Duration.

This commit generates the message file in the configuration phase of
cmake while making sure that in is be installed correctly.

The ros1_bridge expects the ROS1 type to be "duration" and not
"std_msgs/Duration"
See https://answers.ros.org/question/356714/libros1_bridgeso-undefined-reference-for-time-conversions/ for reference